+++ cogment/datalog_service.py
@@ -19,141 +19,43 @@ import cogment.api.common_pb2 as common_api
 import cogment.api.datalog_pb2 as datalog_api
 
 from cogment.control import TrialState
-from cogment.datalog import DatalogSession
+from cogment.datalog import DatalogSession, LogParams
+from cogment.parameters import TrialParameters
 from cogment.errors import CogmentError
 from cogment.utils import list_versions, logger
 from cogment.session import RecvObservation, RecvAction, RecvMessage, RecvReward
 
 import asyncio
 
-
-class LogParams:
-    """Class representing the parameters of a trial for the datalog service."""
-
-    def __init__(self, cog_settings):
-        self.max_steps = None
-        self.max_inactivity = None
-        self.datalog = None
-        self.environment = None
-        self.nb_actors = None
-
-        self._actor_indexes = None
-        self._raw_params = None
-        self._settings = cog_settings
-
-    def __str__(self):
-        result = f"LogParams: {self._raw_params}"
-        return result
-
-    # Type of serialized data being produced and consumed by this class.
-    # This is dependent on all the underlying protobuf messages used to
-    # serialize/deserialize, and should be incremented if any of them changes in
-    # a backward or forward incompatible way. API 1.0 could be considered a type 0 or 1.
-    # Current dependencies: TrialParams, DatalogParams, EnvironmentParams, ActorParams,
-    #                       TrialConfig, EnvironmentConfig, ActorConfig
-    def get_serialization_type(self):
-        return 2
-
-    def _set(self, raw_params):
-        self._set_from_params(raw_params)
-
-    def serialize(self):
-        if self._raw_params is None:
-            raise CogmentError("Not set, cannot serialize")
-        return self._raw_params.SerializeToString()
-
-    def deserialize(self, raw_string):
-        params = common_api.TrialParams()
-        params.ParseFromString(raw_string)
-        self._set_from_params(params)
-
-    def _set_from_params(self, raw_params):
-        if type(raw_params) != common_api.TrialParams:
-            raise CogmentError(f"Wrong type of parameters provided [{type(raw_params)}]")
-
-        self.max_steps = raw_params.max_steps
-        self.max_inactivity = raw_params.max_inactivity
-
-        self.datalog = {
-            "endpoint": raw_params.datalog.endpoint,
-            "exclude": raw_params.datalog.exclude_fields,
-        }
-
-        self.environment = {
-            "name": raw_params.environment.name,
-            "endpoint": raw_params.environment.endpoint,
-            "implementation": raw_params.environment.implementation,
-        }
-
-        self.nb_actors = len(raw_params.actors)
-        self._actor_indexes = None
-        self._raw_params = raw_params
-
-    def get_trial_config(self):
-        config = None
-        if self._raw_params.HasField("trial_config"):
-            config = self._settings.trial.config_type()
-            config.ParseFromString(self._raw_params.trial_config.content)
-
-        return config
-
-    def get_environment_config(self):
-        config = None
-        if(self._raw_params.environment.HasField("config")):
-            config = self._settings.environment.config_type()
-            config.ParseFromString(self._raw_params.environment.config.content)
-
-        return config
-
-    def get_actor_index(self, actor_name):
-        if self._actor_indexes is not None:
-            return self._actor_indexes.get(actor_name)
-        else:
-            indexes = {}
-            for index, actor in enumerate(self._raw_params.actors):
-                indexes[actor.name] = index
-            self._actor_indexes = indexes
-            return indexes.get(actor_name)
-
-    def get_actor_name(self, actor_index):
-        return self._raw_params.actors[actor_index].name
-
-    def get_actor(self, actor_index):
-        actor = self._raw_params.actors[actor_index]
-
-        actor_config = None
-        a_c = self._settings.actor_classes.__getattribute__(actor.actor_class)
-        if actor.HasField("config"):
-            actor_config = a_c.config_type()
-            actor_config.ParseFromString(actor.config.content)
-
-        actor_data = {
-            "name": actor.name,
-            "actor_class": actor.actor_class,
-            "endpoint": actor.endpoint,
-            "implementation": actor.implementation,
-            "config": actor_config,
-            "action_space": a_c.action_space,
-            "observation_space": a_c.observation_space,
-        }
-
-        return actor_data
+_deprecation_reported = False
 
 
 class LogSample:
     """Class representing a trial sample for the datalog service."""
 
     def __init__(self, params):
+        global _deprecation_reported
         self.tick_id = None
         self.timestamp = None
         self.state = None
         self.events = None
 
         self._raw_sample = None
+        self._actor_indexes = None
 
-        if type(params) != LogParams:
+        self._log_params = None
+        self._parameters = None
+        if type(params) == LogParams:
+            if not _deprecation_reported:
+                logger.warning(f"Deprecated use of LogParams")
+                _deprecation_reported = True
+            self._log_params = params
+            self._nb_actors = params.nb_actors
+        elif type(params) == TrialParameters:
+            self._parameters = params
+            self._nb_actors = len(params.actors)
+        else:
             raise CogmentError(f"Wrong type of params provided [{type(params)}]")
-        self._params = params
 
     def __str__(self):
         result = f"LogSample: {self._raw_sample}"
@@ -197,25 +99,46 @@ class LogSample:
 
         self._raw_sample = sample
 
+    def _get_actor_index(self, actor_name):
+        if self._actor_indexes is None:
+            indexes = {}
+            if self._log_params:
+                for index, actor in enumerate(self._log_params._raw_params.actors):
+                    indexes[actor.name] = index
+            else:
+                for index, actor in enumerate(self._parameters.actors):
+                    indexes[actor.name] = index
+            self._actor_indexes = indexes
+
+        return self._actor_indexes.get(actor_name)
+
     def all_actor_names(self):
-        for index in range(self._params.nb_actors):
-            yield self._params.get_actor_name(index)
+        if self._log_params:
+            for index in range(self._nb_actors):
+                yield self._log_params.get_actor_name(index)
+        else:
+            for actor in self._parameters.actors:
+                yield actor.name
 
     def get_action(self, actor):
         if isinstance(actor, int):
             actor_index = actor
         elif isinstance(actor, str):
-            actor_index = self._params.get_actor_index(actor)
+            actor_index = self._get_actor_index(actor)
         else:
             raise CogmentError(f"Wrong type of actor parameter [{type(actor)}]: must be int or str")
-        if actor_index is None or actor_index < 0 or actor_index >= self._params.nb_actors:
+        if actor_index is None or actor_index < 0 or actor_index >= self._nb_actors:
             raise CogmentError(f"Invalid actor [{actor}] [{actor_index}]")
+
+        if self._log_params:
+            action_space = self._log_params.get_actor(actor_index)["action_space"]()
+        else:
+            action_space = self._parameters.actors[actor_index]._actor_class.action_space()
 
         if len(self._raw_sample.actions) > 0:
             data = self._raw_sample.actions[actor_index]
-            action = self._params.get_actor(actor_index)["action_space"]()
-            action.ParseFromString(data.content)
-            return RecvAction(actor_index, data, action)
+            action_space.ParseFromString(data.content)
+            return RecvAction(actor_index, data, action_space)
         else:
             return None
 
@@ -223,19 +146,23 @@ class LogSample:
         if isinstance(actor, int):
             actor_index = actor
         elif isinstance(actor, str):
-            actor_index = self._params.get_actor_index(actor)
+            actor_index = self._get_actor_index(actor)
         else:
             raise CogmentError(f"Wrong type of actor parameter [{type(actor)}]: must be int or str")
-        if actor_index is None or actor_index < 0 or actor_index >= self._params.nb_actors:
+        if actor_index is None or actor_index < 0 or actor_index >= self._nb_actors:
             raise CogmentError(f"Invalid actor [{actor}] [{actor_index}]")
+
+        if self._log_params:
+            obs_space = self._log_params.get_actor(actor_index)["observation_space"]()
+        else:
+            obs_space = self._parameters.actors[actor_index]._actor_class.observation_space()
 
         if self._raw_sample.HasField("observations"):
             data = self._raw_sample.observations
             obs_index = data.actors_map[actor_index]
             obs_content = data.observations[obs_index]
-            obs = self._params.get_actor(actor_index)["observation_space"]()
-            obs.ParseFromString(obs_content)
-            return RecvObservation(data, obs)
+            obs_space.ParseFromString(obs_content)
+            return RecvObservation(data, obs_space)
         else:
             return None
 
@@ -259,7 +186,7 @@ async def _read_sample(context, session, settings):
 
             elif request.HasField("sample"):
                 trial_ended = (request.sample.info.state == common_api.TrialState.ENDED)
-                sample = LogSample(session.trial_params)
+                sample = LogSample(session.trial_parameters)
                 sample._set(request.sample)
                 session._new_sample(sample)
                 if trial_ended:
@@ -285,7 +212,7 @@ class DatalogServicer(datalog_grpc_api.DatalogSPServicer):
 
     def __init__(self, impl, cog_settings):
         self._impl = impl
-        self.__cog_settings = cog_settings
+        self._cog_settings = cog_settings
         logger.info("Datalog Service started")
 
     # Override
@@ -301,13 +228,13 @@ class DatalogServicer(datalog_grpc_api.DatalogSPServicer):
             if not request.HasField("trial_params"):
                 raise CogmentError(f"Initial data log request for [{trial_id}] does not contain parameters.")
 
-            trial_params = LogParams(self.__cog_settings)
-            trial_params._set(request.trial_params)
+            trial_parameters = TrialParameters(None)
+            trial_parameters._set(self._cog_settings, request.trial_params)
 
-            session = DatalogSession(self._impl, trial_id, user_id, trial_params)
+            session = DatalogSession(self._impl, trial_id, user_id, trial_parameters)
             user_task = session._start_user_task()
 
-            reader_task = asyncio.create_task(_read_sample(context, session, self.__cog_settings))
+            reader_task = asyncio.create_task(_read_sample(context, session, self._cog_settings))
 
             normal_return = await user_task